+++ docs/RELEASE-NOTES.rst
@@ -10,12 +10,13 @@ Added Functionality
     * Multi Cluster
     * CRD
 
-
 Bug Fixes
 ````````````
+* 
 
 Upgrade notes
 ``````````````
+* The service label `as3-pool-member-priorityGroup` should match with the priority group mentioned in the pool members of the configmap. Values of priority group of pool members of a pool and the service label `as3-pool-member-priorityGroup` will be defauls to 0 if not provided.
 
 2.20.0
 -------------